Major refactor
The previous implementation of this project would only work with dynamic
table view controllers. This solution works with both static and dynamic
table views.

The main difference between this commit and the previous one is the
blank state view and the main view are separate views. Previously, the
blank state view was a `UIView` placed on top of the main view with an
alpha of 0.

This solution is a bit more organized in Storyboard Builder but it did
require a protocol to pass information between the blank state view and
the main view.